Created UsersListRow file and conneted the TASKBUTTON with the functionality
