+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "python.pythonPath": "C:\\Python38\\python.exe"
+    "python.pythonPath": ".venv\\Scripts\\python.exe"
 }
