+++ django_react_proj/users-fe/src/components/NewUserModal/NewUserModal.module.css
@@ -9,7 +9,7 @@
     float: right;
 }
 .edit{
-    background-color: lightblue;
+    background-color: black;
     border: 2px;
     color: black;
     padding: 8px 20px;
