+++ django_react_proj/users-fe/src/components/TaskButtonModal/TaskButtonModal.js
@@ -1,49 +1,40 @@
-import React, { useState, Fragment } from "react";
+import React, {useState} from "react";
 import { Modal, ModalHeader, Button, ModalFooter } from "reactstrap";
-import styles from './TaskButtonModal.module.css';
+import styles from "./TaskButtonModal.module.css";
 
 const TaskButton = (props) => {
-  const [modalFlag, setModalFlag] = useState(
-        false
-  )
-
-  const [taskFlag, setTaskFlag] = useState(
-    false
-  )
+  const [modalFlag, setModalFlag] = useState(false);
 
   const toggle = () => {
-    setModalFlag(previous => (
-       !previous
-     ));
+    setModalFlag((previous) => !previous);
   };
-  
-  const handlerDeleteUser = async id => {
-    try {
-      await props.deleteUser(id)
-      toggle() 
-    } catch (error) {
-      console.log(error)
-    } 
-   };
 
-  const toggleTaskFlag = () => {
-    setTaskFlag(previous => (
-        !previous
-      ));
+  const handlerDeleteUser = async (id) => {
+    try {
+      await props.deleteUser(id);
+      toggle();
+    } catch (error) {
+      console.log(error);
+    }
   };
-   
+
+  // const toggleTaskFlag = () => {
+  //   setTaskFlag(previous => (
+  //       !previous
+  //     ));
+  // };
+
   return (
     <>
       <Button
-        color = {taskFlag ? 'success': 'danger'}
-        onClick = {() => toggleTaskFlag()}>
+        color={props.taskFlag ? "success" : "danger"}
+        onClick={() => props.ontoggleTaskFlag()}
+      >
         TASKBUTTON
       </Button>
 
       <Modal isOpen={modalFlag} toggle={toggle}>
-        <ModalHeader toggle={toggle}>
-          Czy chcesz usunąć zadanie?
-        </ModalHeader>
+        <ModalHeader toggle={toggle}>Czy chcesz usunąć zadanie?</ModalHeader>
 
         <ModalFooter>
           <Button type="button" onClick={() => toggle()}>
@@ -63,4 +54,3 @@ const TaskButton = (props) => {
 };
 
 export default TaskButton;
-
